Drop ChannelUIDs from contants class
The ChannelUIDs are not constants, they are different for each Thing. In
fact, the generated code did not even compile because there is no
constructor in ChannelUID that takes a ThingTypeUID and a String.

The only reason no one has noticed that before is because, due to a bug,
they don't get generated when the channels are in groups, so it has never
been used. That's another good reason to remove them.